Code cleanup.
Fixed interpretation of conflicting CLI options, moved config-related options to a subcommand.
Also keeps main help output related to login functionality smaller.
