+++ src/AppConfig.hs
@@ -1,36 +1,31 @@
 {-# LANGUAGE OverloadedStrings          #-}
+{-# LANGUAGE RecordWildCards            #-}
 
 -- | Company / account names / ids
 module AppConfig (
-  defaultConfigFileName
+  getEnvAWSProfile
+, listAppConfigProfiles
+, loadAppConfig
 , mergeAppConfig
 , newAppConfig
-, readAppConfigFile
 , tryReadAppConfig
 , writeAppConfigFile
 ) where
 
 
+import           Args
 import           Control.Bool
 import           Data.Aeson
 import           Data.Aeson.Encode.Pretty
 import qualified Data.ByteString.Lazy as LB
-import           Data.List.NonEmpty as NEL
+import           Data.Foldable
+import           Data.List.NonEmpty (NonEmpty(..))
+import qualified Data.List.NonEmpty as NEL
 import           Data.Maybe
+import qualified Data.Text as T
 import           System.Directory
-import           System.FilePath
+import           System.Environment
 import           Types
-
-
--- | Reads required app config file, error if not found
-readAppConfigFile :: FilePath
-                  -> IO AppConfig
-readAppConfigFile confFp = do
-  ac <- tryReadAppConfig confFp
-  case ac
-    of Just x -> pure x
-       Nothing -> error $ "Config file " <> confFp <>
-                          " was not found or was not accessible, please configure first!"
 
 
 -- | Reads config file if present, returns Nothing otherwise
@@ -71,8 +66,59 @@ newAppConfig :: OktaAWSConfig
 newAppConfig oConf = AppConfig $ oConf :| []
 
 
--- | Constructs user-default config path
-defaultConfigFileName :: IO FilePath
-defaultConfigFileName = do
-  h <- getHomeDirectory
-  return $ h </> ".okta-aws-login.json"
+-- | Reads app config file, either default or whichever is given on the command line
+loadAppConfig :: Args
+              -> IO (NonEmpty OktaAWSConfig)
+loadAppConfig Args{..} = do
+  appConf <- readAppConfigFile argsConfigFile
+  envProf <- getEnvAWSProfile
+
+  let defaultConfiguredProfiles = ocAwsProfile <$> NEL.filter (fromMaybe False . ocDefault) (unAppConfig appConf)
+
+       -- consider profiles in the order of preference
+      selectedProfiles = fromMaybe [] $ listToMaybe $ filter (not . null)
+                           [ argsAwsProfiles
+                           , maybeToList envProf
+                           , defaultConfiguredProfiles
+                           ]
+
+      samlConfPredicate OktaAWSConfig{..} = ocAwsProfile `elem` selectedProfiles
+
+      selectedSamlConfigs = NEL.filter samlConfPredicate (unAppConfig appConf)
+
+  case NEL.nonEmpty selectedSamlConfigs
+    of Nothing -> error $ "Please provide at least one AWS profile or specify default(s) (in the config file or via an AWS_PROFILE environmental variable)." <>
+                          " You can re-run with -l to see the list of configured profiles."
+       Just cs -> return cs
+
+
+-- | Returns configured profiles, with an optional default configured profile
+listAppConfigProfiles :: FilePath
+                      -> IO ([AWSProfile], Maybe AWSProfile)
+listAppConfigProfiles confFilePath = do
+  AppConfig{..} <- readAppConfigFile confFilePath
+
+  let isDefaultConfig = fromMaybe False . ocDefault
+
+      maybeDefaultProfile = ocAwsProfile <$> find isDefaultConfig unAppConfig
+
+      allProfiles = toList $ fmap ocAwsProfile unAppConfig
+
+  return (allProfiles, maybeDefaultProfile)
+
+
+
+-- | Reads required app config file, error if not found
+readAppConfigFile :: FilePath
+                  -> IO AppConfig
+readAppConfigFile confFp = do
+  ac <- tryReadAppConfig confFp
+  case ac
+    of Just x -> pure x
+       Nothing -> error $ "Config file " <> confFp <>
+                          " was not found or was not accessible, please configure first!"
+
+
+-- | Looks up AWS_PROFILE env var
+getEnvAWSProfile :: IO (Maybe AWSProfile)
+getEnvAWSProfile = fmap (AWSProfile . T.pack) <$> lookupEnv "AWS_PROFILE"
