+++ src/Args.hs
@@ -1,24 +1,28 @@
+{-# LANGUAGE TemplateHaskell    #-}
+
 module Args (
   Args(..)
 , Command(..)
 , ConfigArgs(..)
-, runWithCommand
+, parseCLICommand
 ) where
 
 
-import           AppConfig
+import           Data.Functor
 import qualified Data.Text as T
+import           Development.GitRev
 import           Network.AWS.Data
 import           Network.AWS.Prelude (Natural)
 import           Network.AWS.Types
 import           Options.Applicative
+import           System.Directory
+import           System.FilePath
 import qualified Text.PrettyPrint.ANSI.Leijen as PP
 import           Types
 
+
 -- | Arguments used for main (login to AWS) functionality
 data Args = Args { argsVerbose :: !Bool
-                 , argsVersion :: !Bool
-                 , argsListAwsProfiles :: !Bool
                  , argsUserName :: !(Maybe UserName)
                  , argsAwsProfiles :: ![AWSProfile]
                  , argsRegion :: !Region
@@ -40,55 +44,50 @@ data ConfigArgs = ConfigArgs { confArgsProfile :: !AWSProfile
 
 -- | Top-level command, either proceed with login (most likely)
 --   configure this tool (initially)
-data Command = Login Args
-             | Configure ConfigArgs
+data Command = Configure ConfigArgs
+             | ListProfiles FilePath
+             | Login Args
              deriving Show
 
 
 parseArgs :: FilePath
           -> Parser Args
-parseArgs defConf = Args
-     <$> switch
-         ( long "verbose"
-        <> short 'v'
-        <> help "Be verbose.")
-     <*> switch
-         ( long "version"
-        <> short 'V'
-        <> help "Print version and exit.")
-     <*> switch
-         ( long "list-profiles"
-        <> short 'l'
-        <> help "List available AWS profiles and exit.")
-     <*> optional (UserName . T.pack <$> strOption
-         ( long "user"
-        <> short 'u'
-        <> help "User name." ))
-     <*> many (AWSProfile . T.pack <$> strOption
-         ( long "aws-profile"
-        <> short 'p'
-        <> help "AWS profile. Defaults to value of AWS_PROFILE env var, then to default config entry."))
-     <*> option parseRegion
-         ( long "region"
-        <> short 'r'
-        <> value NorthVirginia
-        <> showDefaultWith (T.unpack . toText)
-        <> help "AWS region." )
-     <*> strOption
-         ( long "config-file"
-        <> short 'c'
-        <> value defConf
-        <> showDefaultWith show
-        <> help "Use alternative config file." )
-     <*> switch
-         ( long "keep-reloading"
-        <> short 'k'
-        <> help "Keep reloading session token hourly (that's the max TTL at the moment). This only works well on a trusted network where you don't need MFA.")
-     <*> optional (
-            not <$> switch (long "no-ecr" <> help "Skip Docker login to ECR registry, default is in the config.")
-            <|>
-            switch (long "ecr" <> help "Attempt Docker login to ECR registry, default is in the config.")
-         )
+parseArgs defConf =
+  (infoOption ("Version: " <> $(gitBranch) <> "@" <> $(gitHash))
+              (long "version" <> short 'V' <> help "Print version and exit.") <*> pure Args)
+  <*> switch
+      ( long "verbose"
+    <> short 'v'
+    <> help "Be verbose.")
+  <*> optional (UserName . T.pack <$> strOption
+      ( long "user"
+    <> short 'u'
+    <> help "User name." ))
+  <*> many (AWSProfile . T.pack <$> strOption
+      ( long "aws-profile"
+    <> short 'p'
+    <> help "AWS profile. Defaults to value of AWS_PROFILE env var, then to default config entry."))
+  <*> option parseRegion
+      ( long "region"
+    <> short 'r'
+    <> value NorthVirginia
+    <> showDefaultWith (T.unpack . toText)
+    <> help "AWS region." )
+  <*> strOption
+      ( long "config-file"
+    <> short 'c'
+    <> value defConf
+    <> showDefaultWith show
+    <> help "Use alternative config file." )
+  <*> switch
+      ( long "keep-reloading"
+    <> short 'k'
+    <> help "Keep reloading session token hourly (that's the max TTL at the moment). This only works well on a trusted network where you don't need MFA.")
+  <*> (
+        flag Nothing (Just False) (long "no-ecr" <> help "Skip Docker login to ECR registry, default is in the config.")
+        <|>
+        flag Nothing (Just True) (long "ecr" <> help "Attempt Docker login to ECR registry, default is in the config.")
+      )
 
 
 parseRegion :: ReadM Region
@@ -98,48 +97,60 @@ parseRegion = eitherReader (fromText . T.pack)
 parseConfigArgs :: FilePath
                 -> Parser ConfigArgs
 parseConfigArgs defConf = ConfigArgs
-      <$> (AWSProfile . T.pack <$> strOption
-           ( long "aws-profile"
-          <> short 'p'
-          <> help "Name of the associated AWS profile."))
-      <*> (OktaEmbedLink . T.pack <$> strOption
-           ( long "okta-embed-link"
-          <> short 'l'
-          <> help "Okta AWS app 'embed' link, ask your Okta administrator."))
-      <*> optional ( switch
-            (long "default"
-           <> short 'd'
-           <> help "User this profile by default."))
-      <*> optional (switch
-            (long "ecr"
-           <> short 'r'
-           <> help "Enable Docker login to ECR registry by default."))
-     <*> strOption
-         ( long "config-file"
-        <> short 'c'
-        <> value defConf
-        <> showDefaultWith show
-        <> help "Use alternative config file." )
-     <*> option auto
-           ( long "session-duration"
-          <> short 's'
-          <> help "STS session duration, seconds. You can provide a value from 900 seconds (15 minutes) up to the maximum session duration setting for the role. Please coordinate with your AWS administrator."
-           )
+  <$> (AWSProfile . T.pack <$> strOption
+        ( long "aws-profile"
+      <> short 'p'
+      <> help "Name of the associated AWS profile."))
+  <*> (OktaEmbedLink . T.pack <$> strOption
+        ( long "okta-embed-link"
+      <> short 'l'
+      <> help "Okta AWS app 'embed' link, ask your Okta administrator."))
+  <*> optional ( switch
+        (long "default"
+        <> short 'd'
+        <> help "User this profile by default."))
+  <*> optional (switch
+        (long "ecr"
+        <> short 'r'
+        <> help "Enable Docker login to ECR registry by default."))
+  <*> strOption
+      ( long "config-file"
+    <> short 'c'
+    <> value defConf
+    <> showDefaultWith show
+    <> help "Use alternative config file." )
+  <*> option auto
+        ( long "session-duration"
+      <> short 's'
+      <> help "STS session duration, seconds. You can provide a value from 900 seconds (15 minutes) up to the maximum session duration setting for the role. Please coordinate with your AWS administrator."
+        )
 
 
 parseCommand :: FilePath -- ^ default config file
              -> Parser Command
 parseCommand defConf =
-  ( switch (long "configure" <> help "Configure AWS profile given an Okta 'embed' link")
-    *> (Configure <$> parseConfigArgs defConf)
-  ) <|> (Login <$> parseArgs defConf)
+  hsubparser (command "configure"
+             (info (Configure <$> parseConfigArgs defConf)
+                   (progDesc "Configure AWS profile given an Okta 'embed' link")))
+  <|>
+  Login <$> parseArgs defConf
+  <|>
+  ( switch ( long "list-profiles" <> short 'l' <> help "List available AWS profiles and exit.")
+    $> ListProfiles defConf
+  )
 
 
-runWithCommand :: (Command -> IO ())
-               -> IO ()
-runWithCommand rwa = do
+-- | Constructs user-default config path
+defaultConfigFileName :: IO FilePath
+defaultConfigFileName = do
+  h <- getHomeDirectory
+  return $ h </> ".okta-aws-login.json"
+
+
+parseCLICommand :: IO Command
+parseCLICommand = do
   defConfFile <- defaultConfigFileName
-  execParser (opts defConfFile) >>= rwa
+  execParser (opts defConfFile)
   where
     opts defConf = info (helper <*> parseCommand defConf)
       ( fullDesc
